# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_interfaces_http_handlers.CreateUserRequest:
    properties:
      date_of_birth:
        type: string
      email:
        maxLength: 100
        type: string
      first_name:
        maxLength: 50
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/loviary_app_backend_internal_domain_shared.Gender'
        enum:
        - male
        - female
        - other
        - prefer_not
      language:
        maxLength: 10
        type: string
      last_name:
        maxLength: 50
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  internal_interfaces_http_handlers.ErrorResponse:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      message:
        example: validation failed
        type: string
    type: object
...
  internal_interfaces_http_handlers.StreakListResponse:
    properties:
      count:
        example: 2
        type: integer
      streaks:
        items:
          $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.StreakResponse'
        type: array
    type: object
  internal_interfaces_http_handlers.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  internal_interfaces_http_handlers.UpdateDeviceRequest:
    properties:
      device_name:
        type: string
      fcm_token:
        type: string
      platform:
        enum:
        - ios
        - android
        - web
        type: string
    required:
    - fcm_token
    - platform
    type: object
...
  internal_interfaces_http_handlers.UpdateUserRequest:
    properties:
      avatar_url:
        type: string
      date_of_birth:
        type: string
      first_name:
        maxLength: 50
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/loviary_app_backend_internal_domain_shared.Gender'
        enum:
        - male
        - female
        - other
        - prefer_not
      language:
        maxLength: 10
        type: string
      last_name:
        maxLength: 50
        type: string
    type: object
...
  loviary_app_backend_internal_domain_shared.Gender:
    enum:
    - male
    - female
    - other
    - prefer_not
    type: string
    x-enum-varnames:
    - GenderMale
    - GenderFemale
    - GenderOther
    - GenderPreferNot
...
  loviary_app_backend_internal_domain_streaks.StreakStatus:
    enum:
    - none
    - active
    - at_risk
    - broken
    - reset
    type: string
    x-enum-varnames:
    - StreakStatusNone
    - StreakStatusActive
    - StreakStatusAtRisk
    - StreakStatusBroken
    - StreakStatusReset
...
  loviary_app_backend_internal_interfaces_http_dto.StreakResponse:
    properties:
      activity_type:
        type: string
      created_at:
        type: string
      current_streak:
        type: integer
      id:
        type: string
      last_active_date:
        type: string
      longest_streak:
        type: integer
      status:
        $ref: '#/definitions/loviary_app_backend_internal_domain_streaks.StreakStatus'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  loviary_app_backend_internal_interfaces_http_dto.UserResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      gender:
        $ref: '#/definitions/loviary_app_backend_internal_domain_shared.Gender'
      id:
        type: string
      is_active:
        type: boolean
      key_couple:
        type: string
      language:
        type: string
      last_name:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@loviary.app
    name: API Support
  description: Loviary - Couple Relationship Tracking API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Loviary API
  version: "1.0"
paths:
...
  /streaks/{activity_type}:
    get:
      consumes:
      - application/json
      description: Get streak for a specific activity type
      parameters:
      - description: Activity type (e.g., date_night, exercise)
        in: path
        name: activity_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Streak data
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.StreakResponse'
        "400":
          description: Activity type required
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: Streak not found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get streak
      tags:
      - streaks
  /streaks/{activity_type}/record:
    post:
      consumes:
      - application/json
      description: Record an activity occurrence and update streak counters
      parameters:
      - description: Activity type (e.g., date_night, exercise)
        in: path
        name: activity_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated streak
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.StreakResponse'
        "400":
          description: Activity type required
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Record activity
      tags:
      - streaks
  /streaks/me:
    get:
      consumes:
      - application/json
      description: Get all activity streaks for the current user
      produces:
      - application/json
      responses:
        "200":
          description: List of streaks with count
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.StreakListResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all streaks
      tags:
      - streaks
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get user information by user ID
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.UserResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update user profile information by ID
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user data
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.UserResponse'
        "400":
          description: Invalid input or user ID
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: Get current authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.UserResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update current authenticated user's profile
      parameters:
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user data
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.UserResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update my profile
      tags:
      - users
  /users/me/device:
    put:
      consumes:
      - application/json
      description: Update device information (FCM token, platform) for push notifications
      parameters:
      - description: Device info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.UpdateDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Device updated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update my device
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user account (alternative endpoint)
      parameters:
      - description: User creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/loviary_app_backend_internal_interfaces_http_dto.UserResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "409":
          description: Duplicate email or username
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_interfaces_http_handlers.ErrorResponse'
      summary: Create user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"